Rename user detail getters
