+++ examples/multi-cursors/src/utils/getName.ts
